+++ lottery-data/src/main/java/de/alpharogroup/lottery/statistics/LotteryNumberCount.java
@@ -1,0 +1,51 @@
+/**
+ * Commercial License
+ *
+ * Copyright (C) 2015 Asterios Raptis - All Rights Reserved
+ *
+ * Proprietary and confidential
+ *
+ * Unauthorized copying of this software and its files,
+ * via any medium is strictly prohibited
+ *
+ * Written by Asterios Raptis
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package de.alpharogroup.lottery.statistics;
+
+import java.util.Map;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.FieldDefaults;
+
+/**
+ * The class {@link LotteryNumberCount} is used to count the lottery numbers in several draws and
+ * can be used for addition or subtraction
+ */
+@Getter
+@Setter
+@EqualsAndHashCode
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder(toBuilder = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class LotteryNumberCount
+{
+	String lotteryGameType;
+	Map<Integer, Integer> numberCounterMap;
+}