+++ CHANGELOG.md
@@ -4,20 +4,27 @@
 Version 5.9-SNAPSHOT
 -------------
 
+ADDED:
+ 
+- gradle as build system
 
+CHANGED:
+
+- removed maven related files
+- changed project nature from maven to gradle nature
 
 Version 5.8
 -------------
 
 ADDED:
  
-- new dependency of silly-beans in version 1 added
+- new dependency of silly-beans in minor version 1.1 added
 
 CHANGED:
 
 - update of parent version to 5.4
-- update of dependency silly-collections version to 5.8
-- update of dependency randomizer version to 6.5
+- update of dependency silly-collections major version to 8
+- update of dependency randomizer version to 6.6
 - moved class-diagramm to test resources
 
 Version 5.7