+++ CHANGELOG.md
@@ -16,6 +16,8 @@ CHANGED:
 - update of dependency silly-collections major version to 8
 - update of dependency randomizer version to 6.6
 - update of dependency silly-beans in minor version 1.1 added
+- removed all lombok dependent imports
+- removed lombok dependency
 
 Version 5.8
 -------------